fix: move filtering and sorting into a separate hook
